pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "A Discord bot project"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []